% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/data_2.R
\docType{data}
\name{ancova}
\alias{ancova}
\title{ANCOVA}
\format{
Um data.frame com 40 linhas e três colunas
\describe{
  \item{\code{Raiz (mm)}}{Tamanho da raiz medido para cada indivíduo}
  \item{\code{Biomassa (g)}}{Peso dos frutos produzidos por cada indivíduo}
  \item{\code{Herbivoria}}{Tratamentos considerando a herbivoria (Sem_herb ou Com_herb)}
}

Um data.frame com 40 linhas e três colunas
\describe{
  \item{\code{Raiz (mm)}}{Tamanho da raiz medido para cada indivíduo}
  \item{\code{Biomassa (g)}}{Peso dos frutos produzidos por cada indivíduo}
  \item{\code{Herbivoria}}{Tratamentos considerando a herbivoria (Sem_herb ou Com_herb)}
}
}
\usage{
ancova

ancova
}
\description{
Dados simulados do efeito da herbivoria na biomassa dos frutos de uma espécie de árvore na Mata Atlântica. O delineamento experimental permitiu que alguns indivíduos sofressem herbivoria e outros não. Também foram medidos o tamanho da raiz dos indivíduos para inseri-la como uma covariável no modelo.

Dados simulados do efeito da herbivoria na biomassa dos frutos de uma espécie de árvore na Mata Atlântica. O delineamento experimental permitiu que alguns indivíduos sofressem herbivoria e outros não. Também foram medidos o tamanho da raiz dos indivíduos para inseri-la como uma covariável no modelo.
}
\details{
ancova
}
\note{
Dados simulados.
}
\references{
Dados simulados
}
\keyword{datasets}
