% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anova_bloco}
\alias{anova_bloco}
\title{ANOVA em bloco}
\format{
Um data.frame com 30 linhas e três colunas
\describe{
  \item{\code{Riqueza}}{Número de espécies de anuros em cada poça artificial}
  \item{\code{Blocos}}{Seis fragmentos florestais}
  \item{\code{Pocas}}{Tratamentos: (Int-100m, Int-50m, borda, Mat-50m ou Mat-100m)}
}
}
\usage{
anova_bloco
}
\description{
Riqueza de espécies de anuros amostradas em poças artificiais instaladas a diferentes distâncias de seis fragmentos florestais no sudeste do Brasil. Os fragmentos florestais apresentam diferenças entre si que não são do interesse do pesquisador. Por isso, eles foram incluídos como blocos nas análises. As poças artificiais foram instaladas em todos os fragmentos florestais com base no seguinte delineamento experimental: i) quatro poças no interior do fragmento a 100 m de distância da borda do fragmento; ii) quatro poças no interior no fragmento a 50 m de distância da borda do fragmento; iii) quatro poças na borda do fragmento; iv) quatro poças na matriz de pastagem a 50 m de distância da borda do fragmento; e v) quatro poças na matriz de pastagem a 100 m de distância da borda do fragmento. Percebam que todos os tratamentos foram instalados em todos os blocos.
}
\details{
anova_bloco
}
\references{
da Silva, F. R., Oliveira,T. A. L.,  Gibbs, J. P. &  Rossa-Feres, D. C. (2011). An experimental assessment of landscape configuration effects on frog and toad abundance and diversity in tropical agro-savannah landscapes of southeastern Brazil.” Landscape Ecology 27 (1), 87–96. https://doi.org/10.1007/s10980-011-9670-7.
}
\keyword{datasets}
