% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anova_dois_fatores_interacao1}
\alias{anova_dois_fatores_interacao1}
\title{ANOVA de dois fatores com interação}
\format{
Um data.frame com 40 linhas e três colunas
\describe{
  \item{\code{Tempo (minutos)}}{Tempo para eliminar a droga do corpo}
  \item{\code{Pessoas}}{Tratamentos: (XX ou XY)}
  \item{\code{Idade}}{Tratamentos: (Jovem ou Idoso)}
}
}
\usage{
anova_dois_fatores_interacao1
}
\description{
Dados simulados de um experimento com 40 pacientes distribuídos da seguinte maneira: i) 10 indivíduos XX - jovens, ii) 10 indivíduos XX - idosas, iii) 10 indivíduos XY - jovens, e iv) 10 indivíduos XY - idosos. A planilha contém informações sobre o tempo que o corpo leva para eliminar uma droga utilizada em exames de ressonância magnética.
}
\details{
anova_dois_fatores_interacao1
}
\note{
Dados simulados.
}
\keyword{datasets}
