% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anova_simples}
\alias{anova_simples}
\title{Anova simples}
\format{
Um data.frame com 36 linhas e duas colunas
\describe{
  \item{\code{Crescimento (m)}}{Tamanho médio da planta}
  \item{\code{tratamento)}}{Tratamento recebido (Controle ou Adubo_Tradicional)}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
anova_simples
}
\description{
Dados experimentais de crescimento de plantas, XXX etc...
}
\references{
R Core Team (2021). R: A language and environment for statistical computing. R
Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/.
}
\keyword{datasets}
