% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\name{barbaro2009a_amb}
\alias{barbaro2009a_amb}
\title{barbaro2009a_amb}
\format{

}
\description{
Características ambientais (p. ex.: elevação, índice de aridez) das localidades de ocorrência de 36 espécies de plantas
}
\references{
Barbaro, L., & van Halder, I. (2009). Linking bird, carabid beetle and butterfly life‐history traits to habitat fragmentation in mosaic landscapes. Ecography, 32(2), 321–333. doi:10.1111/j.1600-0587.2008.05546.x
}
