% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\name{barbaro2009a_trait}
\alias{barbaro2009a_trait}
\title{barbaro2009a_trait}
\format{
Um data frame com 36 linhas (espécies) e 12 colunas (atributos):
\describe{
\item{\code{trend}}{tendência populacional: (1) aumentando, (2) estável, (3) decrescendo}
\item{\code{redlist}}{raridade na europa: (1) não-ameaçada, (2) ameaçada}
\item{\code{regio}}{raridade regional: (1) >15 dados regionais, (2) 10-15 dados regionais, (3) 4-9 dados regionais, (4) <3 dados regionais}
\item{\code{biog}}{posição biogeográfica: (1) mediterrâneo ou atlântico, (2) amplamente distribuída, (3) norte ou central}
\item{\code{activ}}{atividade diária: (1) diurna, (2) diurna/noturna, (3) noturna}
\item{\code{diet}}{tipo de dieta: (1) collembola, (2) predador generalista, (3) fitófago ou misto}
\item{\code{winter}}{hibernação: (1) somente imago, (2) imago e larva}
\item{\code{color}}{cor corporal: (1) preto ou marrom, (2) metálico}
\item{\code{breed}}{estação de reprodução: (1) primavera, (2) verão, (3) outono}
\item{\code{body}}{tamanho corporal: (1) <6mm, (2) 6-7,9mm, (3) 8-9,9mm, (4) 10-11,9mm}
\item{\code{wing}}{desenvolvimento da asa: (1) braquíptero, (2) dimórfico, (3) macroptero}
\item{\code{period}}{período de atividade do adulto: (1) início da primavera, (2) final da primavera, (3) verão/outono}
}
}
\description{
Atributos funcionais de 36 espécies de besouros: raridade, posição biogeográfica, atividade diária
}
\references{
Barbaro, L., & van Halder, I. (2009). Linking bird, carabid beetle and butterfly life‐history traits to habitat fragmentation in mosaic landscapes. Ecography, 32(2), 321–333. doi:10.1111/j.1600-0587.2008.05546.x
}
