% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\docType{data}
\name{comm}
\alias{comm}
\title{Inserção de espécies na filogenia}
\format{
Data frame com 10 linhas e 32 colunas
\describe{
\item{\code{linhas}}{Comunidades}
\item{\code{colunas}}{Espécies de plantas}
}
}
\usage{
comm
}
\description{
Dados simulados da presença e ausência de 32 espécies de plantas em 10 comunidades
}
\references{
Dados simulados
}
\keyword{datasets}
