% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\docType{data}
\name{composicao_anuros_div_taxonomica}
\alias{composicao_anuros_div_taxonomica}
\title{Análises de diversidade taxonômica}
\format{
Um data.frame com 10 linhas e 10 colunas
\describe{
\item{\code{Linhas}}{comunidades}
\item{\code{Colunas}}{espécies}
}
}
\usage{
composicao_anuros_div_taxonomica
}
\description{
Dados simulados de abundância das espécies registradas em 10 comunidades.
}
\references{
Dados simulados
}
\keyword{datasets}
