% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\docType{data}
\name{composicao_aves_filogenetica}
\alias{composicao_aves_filogenetica}
\title{Composição de aves}
\format{
Data frame com 10 linhas e 37 colunas
\describe{
\item{\code{linhas}}{Comunidades}
\item{\code{colunas}}{Espécies de aves}
}
}
\usage{
composicao_aves_filogenetica
}
\description{
Dados simulados da abundância de 37 espécies de aves registradas em 10 comunidades de floresta atlântica
}
\references{
Dados simulados
}
\keyword{datasets}
