% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/data_2.R
\docType{data}
\name{correlacao}
\alias{correlacao}
\title{Correlação}
\format{
Um data.frame com 35 linhas e duas colunas
\describe{
  \item{\code{Tamanho_raiz}}{Tamanho da raiz de cada indivíduo}
  \item{\code{Tamanho_tronco}}{Tamanho do tronco de cada indivíduo}
}

Um data.frame com 35 linhas e duas colunas
\describe{
  \item{\code{Tamanho_raiz}}{Tamanho da raiz de cada indivíduo}
  \item{\code{Tamanho_tronco}}{Tamanho do tronco de cada indivíduo}
}
}
\usage{
correlacao

correlacao
}
\description{
Dados simulados da altura do tronco e o tamanho da raiz medidos em 35 indivíduos de uma espécie vegetal arbustiva.

Dados simulados da altura do tronco e o tamanho da raiz medidos em 35 indivíduos de uma espécie vegetal arbustiva.
}
\details{
correlacao
}
\references{
Dados simulados

Dados simulados
}
\keyword{datasets}
