% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\docType{data}
\name{filogenia_aves}
\alias{filogenia_aves}
\title{Filogenia aves}
\format{
Árvore no formato phylo
\describe{
\item{\code{Tips}}{37 espécies de aves da Mata Atlântica}
}
}
\usage{
filogenia_aves
}
\description{
Filogenia ultramétrica com 37 espécies de aves endêmicas da Mata Atlântica.
}
\references{
Jetz, W., Thomas, G. H.,  Joy, J. B.,  Hartmann, K., &  Mooers, A. O. (2012). The global diversity of birds in space and time. Nature, 491, 444–48. https://doi.org/10.1038/nature11631.
}
\keyword{datasets}
