% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\name{frenette2012a_amb}
\alias{frenette2012a_amb}
\title{frenette2012a_amb}
\format{
Um data frame com 55 linhas e 5 colunas (variáveis):
\describe{
\item{\code{Grazing}}{Tratamento com dois níveis: com pastejo (grazed) e sem pastejo (ungrazed)}
\item{\code{Aridity}}{Índice de aridez: evapotranspiração potencial/precipitação}
\item{\code{Exclosure}}{duração da exclusão do pastejo}
\item{\code{Bare.ground.2009}}{\% do solo sem cobertura vegetal em 2009}
\item{\code{Elevation (masl)}}{elevação em metro acima do nível do mar}
}
}
\description{
Características ambientais das localidades de ocorrência de 32 espécies de plantas
}
\references{
Frenette-Dussault, C., Shipley, B., Léger, J.-F., Meziane, D., & Hingrat, Y. (2012). Functional structure of an arid steppe plant community reveals similarities with Grime’s       'C-S-R theory. Journal of Vegetation Science, 23(2), 208–222.                                              doi:10.1111/j.1654-1103.2011.01350.x
}
