% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\name{frenette2012a_comu}
\alias{frenette2012a_comu}
\title{frenette2012a_comu}
\format{
Um data frame com 101 linhas e 32 colunas (espécies):
\describe{
  \item{\code{spn}}{lista com cada espécie codificada de sp1 a sp32}
}
}
\description{
Dados de abundância de 32 espécies de plantas distribuídas em estepes áridos do leste de Marrocos
}
\references{
Frenette-Dussault, C., Shipley, B., Léger, J.-F., Meziane, D., & Hingrat, Y. (2012).              Functional structure of an arid steppe plant community reveals similarities with Grime’s      C-S-R theory. Journal of Vegetation Science, 23(2), 208–222.                                           doi:10.1111/j.1654-1103.2011.01350.x
}
