% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\name{frenette2012a_trait}
\alias{frenette2012a_trait}
\title{frenette2012a_trait}
\format{
Um data frame com 35 linhas (espécie) e 5 colunas (atributos):
\describe{
  \item{\code{LA}}{área foliar (mm2)}
  \item{\code{SLA}}{área foliar específica (mm2 mg-1)}
  \item{\code{LDMC}}{conteúdo de matéria seca foliar (mg g-1)}
  \item{\code{LN15}}{conteúdo de nitrôgenio 15 foliar (‰)}
  \item{\code{LCC}}{conteúdo de carbono foliar(μg mg-1)}
}
}
\description{
Atributos funcionais de 32 espécies de plantas: forma de vida, modo de dispersão, peso da semente
}
\references{
Frenette-Dussault, C., Shipley, B., Léger, J.-F., Meziane, D., & Hingrat, Y. (2012).              Functional structure of an arid steppe plant community reveals similarities with Grime’s      C-S-R theory. Journal of Vegetation Science, 23(2), 208–222.                                           doi:10.1111/j.1654-1103.2011.01350.x
}
