% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\docType{data}
\name{lagartos}
\alias{lagartos}
\title{Autotomia da cauda de lagarto}
\format{
Data frame com 223 linhas e 6 variáveis
\describe{
\item{\code{Numero}}{Número do indivíduo}
\item{\code{Sex}}{Sexo do lagarto}
\item{\code{SVL}}{Snout-vent length, Comprimento rostro-cloacal, uma medida de tamanho de corpo}
\item{\code{Intact_tail_length}}{Comprimento da cauda intacta}
\item{\code{Autotomized_tail_length}}{Comprimeiro da cauda que foi autotomizada}
\item{\code{Tail_state}}{Estado da cauda, 0 = Autotomizada; 1 = cauda intacta}
}
}
\source{
\url{https://www.tandfonline.com/doi/abs/10.1080/00222933.2020.1840641}
}
\usage{
lagartos
}
\description{
Dados sobre comprimento do corpo, cauda, estado da cauda, e sexo de indivíduos do lagarto Coleodactylus meridionalis.
}
\references{
Camila N. Oliveira, Ikaro H. M. P. Campos, Diogo B. Provete, Míriam C. Guarnieri & Samuel #'C. Ribeiro (2020)Defensive behaviour and tail autotomy in Coleodactylus meridionalis #'(Squamata: Sphaerodactylidae), Journal of Natural History, 54:33-34, 2209-2218, DOI: #'10.1080/00222933.2020.1840641
}
\keyword{datasets}
