% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\docType{data}
\name{morcegos_rarefacao_amostras}
\alias{morcegos_rarefacao_amostras}
\title{Rarefação por amostras}
\format{
Um data.frame com 20 linhas e três colunas
\describe{
  \item{\code{Primeira linha}}{Número de amostragens em cada localidade}
  \item{\code{MC_Tenentes}}{Frequência de ocorrência das espécies de morcegos registradas no fragmento de mata ciliar do córrego dos tenentes}
  \item{\code{MC_Talhadinho}}{Frequência de ocorrência das espécies de morcegos registradas no fragmento de mata ciliar do córrego talhadinho}
  \item{\code{FF_Experimental}}{Frequência de ocorrência das espécies de morcegos registradas no fragmento da fazenda experimental de Pindorama}
}
}
\usage{
morcegos_rarefacao_amostras
}
\description{
Frequência de registros das espécies de morcegos amostradas em três fragmentos florestais: i) Mata Ciliar do Córrego Talhadinho com 12 hectares; ii) Mata Ciliar do Córrego dos Tenentes com 10 hectares; e iii) Fazenda Experimental de Pindorama com 128 hectares.
}
\references{
Breviglieri, C. P. B. (2008). Diversidade de morcegos (Chiroptera; Mammalia) em três áreas do noroeste paulista, com ênfase nas relações tróficas em Phyllostomidae. Universidade Estadual Paulista. https://repositorio.unesp.br/handle/11449/87596
}
\keyword{datasets}
