% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\docType{data}
\name{parasitas}
\alias{parasitas}
\title{Parasitas de lagartos}
\format{
Data frame com 195 linhas e 5 variáveis:
\describe{
  \item{\code{Especie}}{Espécie do lagarto coletado: Phyllopezus_pollicaris e Hemidactylus agrius}
  \item{\code{Sexo}}{Sexo do lagarto coletado}
  \item{\code{CRC}}{Comprimento Rostro-Cloacal (mm) do lagarto coletado}
  \item{\code{Raillietiella_mottae}}{Número de parasitas da espécie Raillietiella_mottae encontrados}
}
}
\source{
\url{https://www.sciendo.com/article/10.2478/helm-2018-0006}
}
\usage{
parasitas
}
\description{
Dados de morfometria de espécies de lagartos e quantidade de parasitas
}
\references{
Lima, V. F., Brito, S. V., Araujo Filho, J. A., Teles, D. A., Ribeiro, S. C., Teixeira, A. A. M., Pereira, A. M. A. and Almeida, W. O.. " (Pentastomida: Raillietiellidae) parasitizing four species of Gekkota lizards (Gekkonidae and Phyllodactylidae) in the Brazilian Caatinga" Helminthologia, vol.55, no.2, 2018, pp.140-145. https://doi.org/10.2478/helm-2018-0006
}
\keyword{datasets}
