% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\docType{data}
\name{poca_anuros}
\alias{poca_anuros}
\title{Estimadores de riqueza}
\format{
Um data.frame com 14 linhas e 17 colunas
\describe{
  \item{\code{Linhas}}{cada dia de amostragem no habitat reprodutivo}
  \item{\code{Colunas}}{17 espécies de anuros registradas nas amostragens}
}
}
\usage{
poca_anuros
}
\description{
Abundância de 17 espécies de anuros amostradas em 14 dias de coletas de campo em um habitat reprodutivo localizado na região noroeste do estado de São Paulo, Brasil.
}
\references{
da Silva, F. R. Dados não publicados.
}
\keyword{datasets}
