% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\docType{data}
\name{precipitacao_div_taxonomica}
\alias{precipitacao_div_taxonomica}
\title{Dados de precipitação anual nas comunidades}
\format{
Um data.frame com 10 linhas e 1 coluna
\describe{
\item{\code{Linhas}}{comunidades}
\item{\code{prec (mm)}}{Precipitação anual}
}
}
\usage{
precipitacao_div_taxonomica
}
\description{
Dados simulados de precipitação anual (mm) de 10 localidades.
}
\references{
Dados simulados
}
\keyword{datasets}
