% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\docType{data}
\name{precipitacao_filogenetica}
\alias{precipitacao_filogenetica}
\title{Precipitação anual nas comunidades}
\format{
Data frame com 10 linhas e uma coluna.
\describe{
\item{\code{linhas}}{Comunidades}
\item{\code{prec (mm)}}{Precipitação anual}
}
}
\usage{
precipitacao_filogenetica
}
\description{
Dados simulados da precipitação anual em 10 localidades.
}
\references{
Dados simulados
}
\keyword{datasets}
