% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\docType{data}
\name{rarefacao_anuros}
\alias{rarefacao_anuros}
\title{Rarefação dados anuros}
\format{
Um data.frame com 27 linhas e duas colunas
\describe{
\item{\code{Coleta Campo}}{Abundância das espécies de anuros registradas nas coletas de campo}
\item{\code{Colecoes Cientificas}}{Abundância das espécies de anuros registradas nas coleções científicas}
}
}
\usage{
rarefacao_anuros
}
\description{
Número de espécies de anuros registrados usando informações dos indivíduos depositados em coleções científicas e coletas de campo.
}
\references{
da Silva, F. R.,  Provete, D. B., Gerass, L. K., & Bovo, R. P. (2017). What do data from fieldwork and scientific collections tell us about species richness and composition of amphibians and reptiles? South American Journal of Herpetology, 12 (2), 99–106. https://doi.org/10.2994/sajh-d-16-00051.1.
}
\keyword{datasets}
