% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/data_2.R
\docType{data}
\name{rarefacao_morcegos}
\alias{rarefacao_morcegos}
\title{Rarefação dados morcegos}
\format{
Um data.frame com 24 linhas e três colunas
\describe{
\item{\code{MC_Tenentes}}{Abundância das espécies de morcegos registradas no fragmento de mata ciliar do córrego dos tenentes}
\item{\code{MC_Talhadinho}}{Abundância das espécies de morcegos registradas no fragmento de mata ciliar do córrego talhadinho}
\item{\code{FF_Experimental}}{Abundância das espécies de morcegos registradas no fragmento da fazenda experimental de Pindorama}
}

Um data.frame com 24 linhas e três colunas
\describe{
\item{\code{MC_Tenentes}}{Abundância das espécies de morcegos registradas no fragmento de mata ciliar do córrego dos tenentes}
\item{\code{MC_Talhadinho}}{Abundância das espécies de morcegos registradas no fragmento de mata ciliar do córrego talhadinho}
\item{\code{FF_Experimental}}{Abundância das espécies de morcegos registradas no fragmento da fazenda experimental de Pindorama}
}
}
\usage{
rarefacao_morcegos

rarefacao_morcegos
}
\description{
Abundância das espécies de morcegos amostradas em três fragmentos florestais: i) Mata Ciliar do Córrego Talhadinho com 12 hectares; ii) Mata Ciliar do Córrego dos Tenentes com 10 hectares; e iii) Fazenda Experimental de Pindorama com 128 hectares

Abundância das espécies de morcegos amostradas em três fragmentos florestais: i) Mata Ciliar do Córrego Talhadinho com 12 hectares; ii) Mata Ciliar do Córrego dos Tenentes com 10 hectares; e iii) Fazenda Experimental de Pindorama com 128 hectares
}
\details{
rarefacao_morcegos
}
\references{
Breviglieri, C. P. B. (2008). Diversidade de morcegos (Chiroptera; Mammalia) em três áreas do noroeste paulista, com ênfase nas relações tróficas em Phyllostomidae. Universidade Estadual Paulista. https://repositorio.unesp.br/handle/11449/87596

Breviglieri, C. P. B. (2008). Diversidade de morcegos (Chiroptera; Mammalia) em três áreas do noroeste paulista, com ênfase nas relações tróficas em Phyllostomidae. Universidade Estadual Paulista. https://repositorio.unesp.br/handle/11449/87596
}
\keyword{datasets}
