% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/data_2.R
\docType{data}
\name{regressoes}
\alias{regressoes}
\title{Regressões}
\format{
Um data.frame com 109 linhas e quatro colunas
\describe{
  \item{\code{Municipio}}{Nome dos municípios}
  \item{\code{CRC(milímetros}}{média do comprimento rostro-cloacal dos indivíduos em cada município}
  \item{\code{Temperatura(°C)}}{temperatura média dos municípios}
  \item{\code{Precipitação(milímetros)}}{precipitação anual nos municípios}
}

Um data.frame com 109 linhas e quatro colunas
\describe{
  \item{\code{Municipio}}{Nome dos municípios}
  \item{\code{CRC(milímetros}}{média do comprimento rostro-cloacal dos indivíduos em cada município}
  \item{\code{Temperatura(°C)}}{temperatura média dos municípios}
  \item{\code{Precipitação(milímetros)}}{precipitação anual nos municípios}
}
}
\usage{
regressoes

regressoes
}
\description{
Temperatura média anual, precipitação anual e o tamanho médio do comprimento rostro-cloacal (CRC) de populações de Dendropsophus minutus (Anura:Hylidae) amostradas em 109 localidades no Brasil.

Temperatura média anual, precipitação anual e o tamanho médio do comprimento rostro-cloacal (CRC) de populações de Dendropsophus minutus (Anura:Hylidae) amostradas em 109 localidades no Brasil.
}
\details{
regressoes
}
\references{
Boaratti, A. Z.  & da Silva, F. R. (2015). Relationships between environmental gradients and geographic variation in the intraspecific body size of three species of frogs (Anura).” Austral Ecology, 40 (8), 869–76. https://doi.org/10.1111/aec.1226

Boaratti, A. Z.  & da Silva, F. R. (2015). Relationships between environmental gradients and geographic variation in the intraspecific body size of three species of frogs (Anura).” Austral Ecology, 40 (8), 869–76. https://doi.org/10.1111/aec.1226
}
\keyword{datasets}
