% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\docType{data}
\name{sp_list}
\alias{sp_list}
\title{Inserção de espécies na filogenia}
\format{
Data frame com 32 linhas e 3 colunas
\describe{
\item{\code{species}}{nome científico ou morfoespécies de 32 espécies de plantas}
\item{\code{genus}}{gêneros das espécies de plantas}
\item{\code{family}}{família das espécies de plantas}
}
}
\usage{
sp_list
}
\description{
Dados simulados com espécies, gêneros e famílias de 32 espécies de plantas.
}
\references{
Dados simulados
}
\keyword{datasets}
