% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/data_2.R
\docType{data}
\name{teste_t_pareado}
\alias{teste_t_pareado}
\title{Teste t Pareado}
\format{
Um data.frame com 54 linhas e três colunas
\describe{
  \item{\code{Areas}}{27 localidades amostradas}
  \item{\code{Riqueza}}{Número de espécies de artrópodes em cada localidade}
  \item{\code{Estado}}{Tratamento: Pré-queimada ou Pós-queimada}
}

Um data.frame com 54 linhas e três colunas
\describe{
  \item{\code{Areas}}{27 localidades amostradas}
  \item{\code{Riqueza}}{Número de espécies de artrópodes em cada localidade}
  \item{\code{Estado}}{Tratamento: Pré-queimada ou Pós-queimada}
}
}
\usage{
teste_t_pareado

teste_t_pareado
}
\description{
Dados simulados da riqueza de espécies de artrópodes registradas em 27 localidades. Todas as localidades foram amostradas duas vezes. A primeira amostragem foi realizada na localidade antes da perturbação e a segunda amostragem foi realizada após a localidade ter sofrido uma queimada. Portanto, existe uma dependência temporal uma vez que a mesma localidade foi amostrada antes e depois da queimada.

Dados simulados da riqueza de espécies de artrópodes registradas em 27 localidades. Todas as localidades foram amostradas duas vezes. A primeira amostragem foi realizada na localidade antes da perturbação e a segunda amostragem foi realizada após a localidade ter sofrido uma queimada. Portanto, existe uma dependência temporal uma vez que a mesma localidade foi amostrada antes e depois da queimada.
}
\details{
teste_t_pareado
}
\references{
Dados simulados

Dados simulados
}
\keyword{datasets}
