% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/data_2.R
\docType{data}
\name{teste_t_var_diferente}
\alias{teste_t_var_diferente}
\title{Teste t com variâncias diferentes}
\format{
Um data.frame com 12 linhas e duas colunas
\describe{
\item{\code{CRC (milímetros)}}{Comprimento rostro-cloacal de cada indivíduo de L. podicipinus}
\item{\code{Estacao}}{Tratamentos considerando as estações do ano (Chuvosa ou Seca)}
}

Um data.frame com 12 linhas e duas colunas
\describe{
\item{\code{CRC (milímetros)}}{Comprimento rostro-cloacal de cada indivíduo de L. podicipinus}
\item{\code{Estacao}}{Tratamentos considerando as estações do ano (Chuvosa ou Seca)}
}
}
\usage{
teste_t_var_diferente

teste_t_var_diferente
}
\description{
Comprimento rostro-cloacal (CRC) de fêmeas de Leptodactylus podicipinus amostradas em diferentes estações do ano com armadilhas de interceptação e queda na Região Noroeste do estado de São Paulo

Comprimento rostro-cloacal (CRC) de fêmeas de Leptodactylus podicipinus amostradas em diferentes estações do ano com armadilhas de interceptação e queda na Região Noroeste do estado de São Paulo
}
\details{
teste_t_var_diferente
}
\references{
da Silva, F. R., & Rossa-Feres, D. C. ( 2010). “Seasonal variation in body size of tropical anuran amphibians.” Herpetology Notes, 3, 205–9.

da Silva, F. R., & Rossa-Feres, D. C. ( 2010). “Seasonal variation in body size of tropical anuran amphibians.” Herpetology Notes, 3, 205–9.
}
\keyword{datasets}
