% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teste_t_var_igual}
\alias{teste_t_var_igual}
\title{Teste t com variâncias iguais}
\format{
Um data.frame com 51 linhas e duas colunas
\describe{
  \item{\code{CRC (milímetros)}}{Comprimento rostro-cloacal de cada indivíduo de P. nattereri}
  \item{\code{Estacao}}{Tratamentos considerando as estações do ano (Chuvosa ou Seca)}
}
}
\usage{
teste_t_var_igual
}
\description{
Comprimento rostro-cloacal (CRC) de machos de Physalaemus nattereri (Anura:Leptodactylidae) amostrados em diferentes estações do ano com armadilhas de interceptação e queda na Região Noroeste do estado de São Paulo
}
\details{
test_t_var_igual
}
\references{
da Silva, F. R., & Rossa-Feres, D. C. ( 2010). “Seasonal variation in body size of tropical anuran amphibians.” Herpetology Notes, 3, 205–9.
}
\keyword{datasets}
