% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\docType{data}
\name{uv_cells}
\alias{uv_cells}
\title{Células sanguíneas de girinos}
\format{
Data frame com 50 linhas e 8 variáveis
\describe{
  \item{\code{UV}}{Tratamentos, tempos de exposição à luz Ultravioleta: Controle (CT), 6h, 12h, 18h, 24h}
  \item{\code{Pigmentation}}{Fator com dois níveis: codifica se o girino tinha ou não pigmentação}
  \item{\code{Total_Cell}}{Número total de células contadas}
  \item{\code{Lymphocyte}}{Número de linfócitos encontrados no sangue}
  \item{\code{Neutrophil}}{Número de neutrófilos encontrados no sangue}
  \item{\code{Basophil}}{Número de basófilos encontrados no sangue}
  \item{\code{Monocyte}}{Número de monócitos encontrados no sangue}
  \item{\code{Eosinophil}}{Número de eosinófilos encontrados no sangue}
}
}
\source{
\url{https://www.sciencedirect.com/science/article/pii/S1011134417311727}
}
\usage{
uv_cells
}
\description{
Dados de um experimento que testou como diferentes tempos de exposição à luz ultra-violeta altera as células sanguíneas em girinos de LIthobates catesbeianus com e sem pigmentação.
}
\references{
Lilian Franco-Belussi, Lara Zácari Fanali, Classius De Oliveira, 2018. UV-B affects the immune system and promotes nuclear abnormalities in pigmented and non-pigmented bullfrog tadpoles. Journal of Photochemistry and Photobiology B: Biology, 180, 109-117
}
\keyword{datasets}
