% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/data_2.R
\docType{data}
\name{anova_simples}
\alias{anova_simples}
\title{ANOVA Simples}
\format{
Um data.frame com 36 linhas e duas colunas
\describe{
  \item{\code{Crescimento (mm)}}{Crescimento dos indivíduos durante o período do experimento}
  \item{\code{Tratamento}}{Tratamentos: (Controle, Adubo_Tradicional ou Adubo_X_2020)}
}

Um data.frame com 36 linhas e duas colunas
\describe{
  \item{\code{Crescimento (mm)}}{Crescimento dos indivíduos durante o período do experimento}
  \item{\code{Tratamento}}{Tratamentos: (Controle, Adubo_Tradicional ou Adubo_X_2020)}
}
}
\usage{
anova_simples

anova_simples
}
\description{
Dados simulados de um experimento com indivíduos de Coffea arabica cultivados em três grupos: i) grupo controle onde os indivíduos não receberam adubação, ii) grupo onde os indivíduos receberam a adição do adubo tradicional mais utilizado pelos produtores de C. arabica, e iii) grupo onde os indivíduos receberam a adição do adubo X-2020.

Dados simulados de um experimento com indivíduos de Coffea arabica cultivados em três grupos: i) grupo controle onde os indivíduos não receberam adubação, ii) grupo onde os indivíduos receberam a adição do adubo tradicional mais utilizado pelos produtores de C. arabica, e iii) grupo onde os indivíduos receberam a adição do adubo X-2020.
}
\details{
anova_simples
}
\note{
Dados simulados.
Dados simulados
}
\references{
Dados simulados
}
\keyword{datasets}
