% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2.R
\name{frenette2012a_amb}
\alias{frenette2012a_amb}
\title{frenette2012a_amb}
\format{

}
\description{
Características ambientais das localidades de ocorrência de 32 espécies de plantas
}
\references{
Frenette-Dussault, C., Shipley, B., Léger, J.-F., Meziane, D., & Hingrat, Y. (2012). Functional structure of an arid steppe plant community reveals similarities with Grime’s       'C-S-R theory. Journal of Vegetation Science, 23(2), 208–222.                                              doi:10.1111/j.1654-1103.2011.01350.x
}
